Feautures

+++ src/main/java/com/digitalbanking/bankingproject/dto/AccountsResponseDTO.java
@@ -3,10 +3,12 @@ package com.digitalbanking.bankingproject.dto;
 import com.digitalbanking.bankingproject.constants.AccountStatus;
 import com.digitalbanking.bankingproject.constants.AccountType;
 
+import java.math.BigDecimal;
+
 public record AccountsResponseDTO(
         String currency,
         String iban,
-        Double balance,
+        BigDecimal balance,
         AccountType accountType,
         AccountStatus accountStatus
 ) {
